Change the structure of the whole project
