+++ src/models/reservation.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
-#include "reservation.hpp"
-#include "student.hpp"
-#include "dining_hall.hpp"
-#include "meal.hpp"
+#include "models/reservation.hpp"
+#include "models/student.hpp"
+#include "models/dining_hall.hpp"
+#include "models/meal.hpp"
 
 // Initialize static member
 int Reservation::next_reserve_id = 10001;
