+++ src/models/user.cpp
@@ -1,0 +1,3 @@
+#include "models/user.hpp"
+
+// nothing to add yet
