--- src/user.cpp
@@ -1,3 +1,0 @@
-#include "user.hpp"
-
-// nothing to add yet
